MPEG4Extractor: use Vector<int32_t> encryptedsizes
Use Vector<int32_t> instead of Vector<size_t>
The code below treat encryptedsizes as Vector of 4 bytes
size_t may be 8 bytes on 64-bit platforms.

Bug 148141603

+++ media/extractors/mp4/MPEG4Extractor.cpp
@@ -202,8 +202,8 @@ private:
         uint32_t duration;
         int32_t compositionOffset;
         uint8_t iv[16];
-        Vector<size_t> clearsizes;
-        Vector<size_t> encryptedsizes;
+        Vector<int32_t> clearsizes;
+        Vector<int32_t> encryptedsizes;
     };
     Vector<Sample> mCurrentSamples;
     std::map<off64_t, uint32_t> mDrmOffsets;
